requirements.txt
anthropic>=0.83.0
fastapi>=0.132.0
httpx>=0.28.1
pydantic>=2.12.0
pydantic-settings>=2.13.0
uvicorn>=0.41.0
pyyaml>=6.0
